+++ R/posterior_epred.R
@@ -491,6 +491,12 @@ posterior_epred_zero_inflated_binomial <- function(prep) {
   prep$dpars$mu * trials * (1 - prep$dpars$zi)
 }
 
+posterior_epred_zero_inflated_beta_binomial <- function(prep) {
+  # same as zero_inflated_binom as beta part is included in mu
+  trials <- data2draws(prep$data$trials, dim_mu(prep))
+  prep$dpars$mu * trials * (1 - prep$dpars$zi)
+}
+
 posterior_epred_zero_inflated_beta <- function(prep) {
   with(prep$dpars, mu * (1 - zi)) 
 }